refactor: move Uuid annotation to sub annotation package
